[aim] Cleanup name mapper and tag UUIDs with resource types
The name mapper has been simplified and cleaned up, with the
expectation that deployments will eventually be able to override it
through configuration to customize the mapping between OpenStack and
APIC resource identities. Neutron UUIDs are tagged with their types to
improve debuggability, and reverse mapping functions are provided
where needed.

Change-Id: I7f6a48604a66683ef889f5c804a62dc5b1c6a4c5