[aim-mapping] Update subnetpool for L3P fixes
The ip_pool and subnet_prefix_length were not getting updated
when the subnetpools for L3P were being updated.

The implicit subnetpool was not getting deleted once it was
disassociated from the L3P by an update operation. If the
subnetpool does not have any subnet associations, it will get
deleted. If it has subnet associations it will not get deleted
and a warning will be logged.

If the implicitly created address_scope associated with a L3P
has not subnetpool associations it will get deleted, else a
warning will be logged but the address_scope will not be deleted.

The L3P deletion will proceed in all the above cases.

Change-Id: I08b8d60d2bae0527bf093b412094c9171f885d3a