[AIM] External-connecitvity for unscoped shared networks
Networks that don't use an address-scope can
move between VRFs as they are connected/disconnected
via routers to other networks. Accordingly, the
external-connectivity of the VRFs needs to be
updated. This is required only for the case
where the "router topology" changes VRF - the old
VRF loses external connectivity available through
the router whereas the new VRF gains the same.

Change-Id: I7b6d22405a8c7b0a93a88e7bc186ab7c158757da
Signed-off-by: Amit Bose <amitbose@gmail.com>