[aim-mapping] Fix replacing L3P implicit subnetpool
This patch adds fixes for the following:

1. Replacing the implicitly created subnetpool association
for a L3P with an explicitly created subnetpool where the L3P
is shared and the explicit subnetpool is in a different tenant.
This was earlier broken since elevated context was not being
used for cross tenant access to the explicit subnetpool.

2. Disallow replacing a subnetpool association with L3P
if the subnetpool is in use by way of allocated subnets.

Change-Id: I6b4831eaa209be3ca7be81332d954a5d889e7a2d