[aim] Move port-security unit test
This test was added in the recent commit:
560263d40caf2fbc18e079b89bfee827b93032ae

Although the test passes the tox gate, it fails when using the testtools
test runner. The error seen is that database tables are not present which can be
traced back to incorrect test setup. This test requires both, Neutron, and GBP,
test setup to be configured correctly, and which was not adequately done in the
original commit when the test was added to the test_apic_aim test module. The
better solution is to put this test in the test_aim_mapping test module which
already performs Neutron and GBP setup. By moving the test we avoid polluting the
test_apic_aim module with GBP details.

This patch also reorganizes the GroupPolicyPluginTestCase base class by moving
the test_reverse_on_delete test which had to be overriden at times when this class
was used as a base class but should not have been required to. The test needs
to be only run once for the plugin.

Change-Id: Ib765a5fff4b81b59809c30b2702b8f080fe48a8e