[wip][aim_mapping] APG mapping to AIM ApplicationProfile
Prior to this patch, the same AIM ApplicationProfile (with name 'OpenStack'),
was used for EPGs (a different ApplicationProfile was created in each tenant,
but with the same name).This patch attempts to map the GBP Application
Policy Group (APG) to an AIM ApplicationProfile, such that a new AIM
ApplicationProfile is created for every GBP Application Policy Group with
the name of the ApplicationProfile derived from the ID of the GBP APG. The
ApplicationProfile is created lazily and at the time of creating the first
EPG that needs to be associated with this ApplicationProfile. Correspondingly,
the ApplicationProfile is deleted when the last EPG using this
ApplicationProfile is deleted.

If and when the apic_aim mechanism driver moves an EPG to a new tenant, if
that EPG was earlier in the non-default ApplicationProfile, a new
AplicationProfile is created in the tenant to which the EPG is moved to (and
the ApplicationProfile in the previous tenant is deleted if the EPG being
moved was the last one using the ApplicationProfile with the same name in
that tenant).

Implements: blueprint application-group

Change-Id: I24d12f96d4e15bad5861b00887707fe7cda492a5