[aim_mapping] APG mapping to AIM ApplicationProfile
Prior to this patch, the same AIM ApplicationProfile (with name 'OpenStack'),
was used for EPGs (a different ApplicationProfile was created in each tenant,
but with the same name).This patch attempts to map the GBP Application
Policy Group (APG) to an AIM ApplicationProfile, such that a new AIM
ApplicationProfile is created for every GBP Application Policy Group with
the name of the ApplicationProfile derived from the ID of the GBP APG. The
ApplicationProfile is created lazily and at the time of creating the first
EPG that needs to be associated with this ApplicationProfile. Correspondingly,
the ApplicationProfile is deleted when the last EPG using this
ApplicationProfile is deleted.

Implements: blueprint application-group

Change-Id: I24d12f96d4e15bad5861b00887707fe7cda492a5