Update gbpfunctest to account for auto-ptg - Redux
A test in the gbpfunctest functional test suite occassionally
fails in the aim gate job. This is because it does not take into
consideration the presense of an auto-ptg in addition to a PTG
that it creates. It tries to check if the PTG that it created
is assocaited with the relevant L2P or not by scrapping the output
of the show command. If the created PTG is the first in the displayed
list of PTGs in the L2P show output, the test passes, else it fails.

This patch updates the test logic to account for the additional auto-ptg.

A fix was earlier committed to solve this:
I57b925081273300cfc3e5320af6e05a868d0759e

but it did not fix the issue completely.

Change-Id: Ieeb4ec68bbd6f9f400efad7e2a0b6e477bb5f852