Fix ML2Plus exception handling for Newton
Sync ML2Plus mechanism driver exception handling with changes made to
ML2 in Newton. This should ensure retriable exceptions raised during
precommit for subnetpools and address scopes are retried properly, and
allow propagation of exceptions from mechanism drivers to clients. A
change made to the ScopeUpdateNotSupported exception for Newton that
resulted in logged errors is no longer needed and was reverted.

Change-Id: Ic05b0912a7709a91a90a0ed95b21e63816b8db99