Remove use of verbose flag for logging
This config option is no longer available. Also adds
a method to oslo.db which was present in an older version. These
changes are mainly to faciliate running the tests in the Openstack
gate where stable/newton packages are not being installed for this
branch. These changes should not be backported.

Change-Id: I203d1f9fde4b02116c228806ed8d97c8e63e32ab