[apic-mapping] Ignore FlushError in create PTG
The FlushError happens when create PTG operations happen
concurrently within the same tenant. In the context
of the create PTG post-commit implementation for
apic-mapping policy driver this error is not catastrophic.
In fact, it implies that another concurrent operation has
already succeeded in creating a subnet for this tenant. Hence
we just revert the creation of the subnet, and move on.

Change-Id: If8ab52a884d6fa33ff582f2d6e24c523203d894e
(cherry picked from commit ecf58b648c807f0c92d11cb5ae9f8aab3c931bd6)