Make ensure_tenant retriable
We were eating up all exceptions thrown during the ensure_tenant phase
and throwing a generic exception. In cases where the operation failed
due to concurrent operations, we would like ensure_tenant to be retried.
This adds more conditional logic to the exception handling to facilitate
retries for retriable exceptions.

Change-Id: Idbf6d82ae512b46f35f5c2b9e13b45a4ff8f1d7d