[aim_mapping] Per tenant AIM resrouces cleanup
Certain AIM Filters, FilterEntries, Contracts, ContractSubjects are created
when the first l2_policy is created for a tenant. It is expected that these
be cleaned up when the last l2_policy in that tenant is deleted. The curent
implementation attempts to do this but the logic is not thread-safe in that
when two threads try to delete the last two l2_policies neither ends up
cleaning the AIM resources since each thinks that its not the last l2_policy
in the tenant (the logic is based on retrieving the l2_policy count).

This patch attempts to ensure that the AIM resources' cleanup indeed happens
after the last l2_policy in the tenant is deleted by moving the AIM resource
cleanup out of the l2_policy delete operation's transaction, and performing
it in a separate transaction triggered from a after_commit hook. This
post_commit hook overrides an already existing after_commit hook used for
pushing notifications.

Change-Id: I2297d36ee9946a3d7e0859ced752462b049e4e25