Prioritizing in-process notifications
The aim_mapping policy driver currently leverages a scheme of delaying
dispatching notifications until the outermost transaction, which
initiated the notification, has committed. However, Neutron ML2 plugin
is moving to a model of more in-process notifications, and some of these
changes are being backported to stable/newton as well (see:
https://github.com/openstack/neutron/commit/4a96b962b560733a90b24ae10290e3d835d448e3).
These in-process notifications need not be delayed, and in fact they
should be dispatched and processed immediately, and within the scope
of the transaction.

This patch attempts to identify the notifications that need to be dispatched
immediately versus those that need to be delayed based on the module
names of the notfication handlers. The currently known list of agent and
registry notification handlers (used by the aim_mapping driver) is captured
in the OUT_OF_PROCESS_NOTIFICATIONS list in:
gbpservice/network/neutronv2/local_api.py

If a notification handler does not belong to the above list, it is dispatched
immediately.

As and when additional notification handlers are identified for which the
notification needs to be queued for delayed dispath, they should be added to
the above OUT_OF_PROCESS_NOTIFICATIONS list.

Change-Id: Ib5beb9c8056463c875c775ad19d2d3cbee3430dd