[apic_aim] Multi-scope routing
Allow subnets with different address scopes, as well as unscoped
subnets, to be attached as interfaces to the same router. Note that no
East/West routing is provided between differently scoped interfaces of
a router, but East/West routing is provided within each scope and
North/South routing is provided between each of them and the router's
gateway.

Routed IPv4 and IPv6 subnets on the same network currently either must
both be unscoped or must be associated with isomorphic address scopes
(referencing the same VRF). Adding a subnet to a router results in a
NonIsomorphicNetworkRoutingUnsupported exception if this constaint
would be violated. Eventually, use of identity NAT to move IPv6
traffic from the network's IPv4 VRF to its IPv6 VRF will allow this
constraint to be removed or relaxed.

External connectivity for multi-scope routers may require some
follow-on work to correctly handle all cases.

Change-Id: Idbbd4400e570654937c2bee4577422a91224430e