[wip][aim_mapping] Create implicit AIM contracts for existing l3ps
Uses a configuration driven approach to allow migrating a deployment
which had l3_policies prior to the aim_mapping policy driver moving
to per-l3p implicit AIM contracts (see commit:
https://github.com/openstack/group-based-policy/commit/f50db6f1ce70a29b40b833f224fc539ebb05b107)

The following configuration:
[aim_mapping]
create_per_l3p_implicit_contracts=True

should be set to True and the Neutron server needs to be restarted. The implicit
AIM contracts will be created at the time of the initialization of the AIM
policy driver.

The above configuration can be subsequently either removed or set to False and will
prevent the migration logic to kick in when the next server restart happens.

Since this configuration option is only for migrating newton or prior deployments, it
will be removed in the O release.

Change-Id: I7e5f793bdf3618655600898feba64aac7c099239