[aim] Reduce log level of pymysql exception
The oslo_db code catches the pymysql SAVEPOINT exception, logs it error level,
and then raises it as an oslo_db exception, and due to which the operation
is retried. As such, the ERROR level logging of the pymysql exeption is a
bit misleading. This patch works around it by patching the oslo_db module
to log at debug level instead of error level.

Change-Id: If8afe36706f5d07d7ab9b840c877854b8ed11c42