Speedup py27 with changes to NFP testing
Currently the py27 UT job is consistently running longer than
the gate job definition allows. One of the possible reasons
is the use of the "advanced" mode in the NFP tests initialization
which in turn causes use of RMQ server and which fails with timeouts
and multiple retries. Ideally this RMQ server use should be mocked
for UTs. This patch switches the NFP mode to "base" which does not
use the RMQ server and the timeouts are not being seen anymore.

In addition, this patch also temporarily disables the UTs in the
contrib directory.

Change-Id: I2df99f1b94f5574b15c348df69300692f4e14da3