Disabling UTs for simple_chain_driver
This driver is not used and supported. Since this is a stable branch
we dont want to remove the driver. We are also not deleting the test
module, but just renaming it so that does not get discovered in a tox
run.

We are specifically targeting this test module since it has tests that
invoke the simple_chain_driver code that uses the Heat client to
connect to the Heat server. These connection attempts obviously
fail but they are retried several times, and after timing out on
each occasion. By not running these tests the hope is the py27 UT
job will speed up.

We are also disabling a UT test_policy_target_group_updated from
the test_nfp_node_driver module it takes a long time to complete.

Change-Id: I0170eca7b0248b390beda93e8dbd882daa676e80