[WIP][apic-aim] Data migration for persisting mapping to APIC
A previous commit (https://review.openstack.org/#/c/450309/) added DB
tables mapping Neutron resource identities to APIC resource
identities, but did not include a data migration. This patch populates
the new tables during the DB migration with the the APIC resource
identities for existing Neutron resources, using information from both
the Neutron and AIM DB tables.

This is currently WIP because support for pre-existing address scopes
and for normal and external networks needs to be added.

Change-Id: Ia8a74b9c2289060234716ce89fb4b7b3d1c29596