[apci_aim] Remove use of with_lockmode in ml2plus
In certain cases of concurrent operations we are seeing an error
which suggests that a rolled back transaction is be reused. On debugging
it has been observed that the error manifests when the code path
executes the queries that are using with_lockmode in a couple of
places in the ml2 plugin component. Removing the with_lockmode usage
seems to prevent this issue and does not seem to be affecting the
correctness of behavior even in concurrent execution situations.

This patch removes the use of the with_lockmode in the identified place
when the ml2plus plugin configured.

Change-Id: If65c238cbf49a9cfd2546ca26d37ee721f6f986c