Limiting the number of UTs to run
We have tried several things in the past to speed up the UTs such that
we dont hit the timeout. But it seems that we hit the limit terms of
optimiation, and if we dont get enough worker threads (and which is
happening consistently), the patches keep timinout out and failing
the gate.

This patch creates a split in the unit test directory strucutre by creating
symbolic links. This way, based on the .testr.conf, either GBP or NFP tests
can be run. This is change is intended to be temporary until we figure out a
better solution.

Change-Id: I62a0c010ce4da7f076a51e7be7e0563c72d39a5a