Ocata sync dsvm jobs udpate
This patch updates the setup for the various devstack gate jobs
to make them functional for Ocata and thus also validates
the parent patch Ocata sync patch.

Note that the NFP job still needs to be updated and has been
disabled in this patch. It can be re-enabled whenever its
updated in a followup patch.

Change-Id: I7cd5667fcc86577eb33c0233991cbb687c0ce8c9