Fix delete_subnet patching regression
The change:
I04fb4b323029772bf417a8548ae43e2c93647b65

introduced conditional patching of the delete_subnet
method of the ML2 plugin. On account of this conditional
logic, the delete_subnet method is not patched with RH OSP9,
and consequently the transaction guard decorated version
of the delete_subnet method gets used and prevents
calling the method from within a GBP method's transaction.

This patch enhances the conditional logic by removing
the transaction guard when the delete_subnet method is not
patched.

Change-Id: Ib44f432623a06220ffee8e6bf47568c56f952be9