Fix APICManager Initialization
Commit d612e2a0f98a39c1eae613c0e83be137012efbf6 changed
the initialization of APICManager, which resulted in an
unintentional re-ordering of parameters passed to it. This
caused the apic_system_id parameter to be used incorrectly
in the class, and therefore elements that used the paramter,
such as tenant name in ACI, were using the default value
instead ("openstack").

This patch ensures that the apic_system_id is passed as a
named parameter so that it will be used correctly as part
of initializaiton.

Change-Id: I0e2793c9c4ae8cf7aefd265db7f6ef5307b445f0