Update service_chain heat node driver for LBaaS v2
This patch also updates the chain_mapping policy driver
and the Node Composition Plugin (NCP) to accomodate the
transactional semantics of an async driver such as
the aim_mapping driver which performs all operations
in the pre_commit phase. Note that the chain_mapping driver
and NCP will continue to work with the resource_mapping
driver (or a similar driver) as before.

Since the service chain plugin does not have any knowledge of
pre- or post- commit transaction phases of the GBP plugin that
drive the service chain plugin, a workaround (read hack!) has
been introduced in the current patch to overcome this. Enhancing
the service chain plugin structure is a bigger discussion
and beyond the scope of this patch. It will be tackled in the
context of enhancing and supporting the NFP framework for
aim_mapping like single transaction drivers.

Change-Id: I74f1b3d0d61bd6c3859fcaaf9a569ca3c6cc07ed