Allow policy drivers to add RPC listeners to RPC workers
Neutron provides a framework for associating RPC listeners
with the pool of RPC workers it creates. This framework
is available to the core plugin and service plugins by
implementing the set_rpc_listeners() method. However, there
are cases when the GBP policy drivers might also want to
associate listeners with the RPC workers (as in the case
of the aim_mapping driver). This patch updates the GBP
poilcy driver manager framework to facilitate doing this.

To leverage this framework, the policy driver framework
would need to implement the set_rpc_listeners() method
which returns the RPC listener it instantiated.

Change-Id: I04ff0555c18507b7d9776bf6a26003d894b02a61