Fix DB migrations
Commit 97bdb259907091fc1171122317f3ee9ac6a86172 added a DB
migration to increase the L3P's ip_pool length/size. That patch
breaks the DB migration upgrade path, as the backport patch in the
stable/newton branch would cause some migrations to be missed.

This patch fixes the migration problem by putting it in the
appropriate place to ensure previous migrations won't be skipped.

Change-Id: Ide80f07a519add34c96a16e1462e21aa46bfd94b