Raise OS_TEST_TIMEOUT
Several apic_aim unit tests have been occasionally failing in the
gate, then passing when rechecked. This appears to be due to them
taking longer than 120 seconds to complete. We need to find out why
they take so long, and why the timing varies so much, but this should
help reduce the need to recheck in the mean time.

Change-Id: I6362d1cb6d7df87087282a6b78cabc741b31b9f1