Disable transaction guard on plugin level
Neutron transaction guard throws an exception when certain object
modification apis are called with active session. This happens for
a bunch of GBP flows. Until a better solution is engineered, we
disable the transaction guard.

This solves a bug where network is left behind after ptg is deleted.

Note: this has to be a temporary solution, since the option to disable
transaction guard is temporary in neutron and is expected to go away.

Change-Id: I828ed5d96587831c54d0f6fe3d3ea582f64e184f