Reduce log level for ovo_rpc when using apic_aim
When the aim_mapping policy driver is configured (along with
the apic_aim mechanism driver), a lot of the following messages
are seen in the neutron logs:

This handler is supposed to handle AFTER events, as in 'AFTER it's
committed', not BEFORE. Offending resource event: port, after_update.

The neutron code logs this at error level but allows the execution
to progress. This does not break functionality.

Neutron is right in flagging this, but as far as the apic_aim mechanism
driver is concerned, this behavior is per design since it sends out port
notifications in the pre-commit phase.

Change-Id: Id201f421a676af7c822a182e0ee87f034678e2d7