[apic_aim] Specify ON condition for JOINs
If the ON condition is not specified when calling join() as part of a
query, sqlalchemy infers what columns in the tables to use in the ON
condition of the SQL JOIN. We have encountered at least one bug where
the inferred ON condition did not produce the intended result, and
unexpected data was returned from the query. In order to ensure
similar issues are not going undetected, the ON condition is now
explicitly specified for all JOINs in the apic_aim MD's queries. Doing
so exposed several queries that were including unnecessary JOINs,
which have also been cleaned up.

Change-Id: Ic77a7c7e7b41d27c35f8cf08155d9e1eb1e10d09