Fix DB migration on AIM data
Starting revision cc09261e0fb5, alembic migration transactions
seem to break to the point that the gbp_alembic_version value
in the database after a successful migration end up being 1 or 2
versions behind. The root cause seems to be a missing commit
during the AIM data migration that this patch is adding back.

Change-Id: I22e58454c4294cd502d019cb81ab0c1c176b77f1