Skip logging of sqlalchemy events in UTs
When running most of the UTs, log statements like the following:

"
No sqlalchemy event for resource router found
"

are being seen during the UT teardown stage. In the case of the
py27 gate job it inundates the log and makes it difficult to
find more relevant information.

The log warnings are seen because the unregistering of events is
attempted twice, and the warning is shown on the second instance.
The unregsitering is triggered from the Neutron UT framewor and there
doesn't seem to be much control over this from GBP.

This patch selectively suppresses these messages when running UTs,
and the suppressing behavior is limited to the class and method
where its considered redundant.

Change-Id: I19e83d67b93cfeaa066ef8bd586df8662778dee3