Support AIM VMM/PhysDom Domains
This patch adds support for the recently-added
VMM and PhysDom Domain support in aci-integration-module
(patch 2eadd1d1850fc397d79823ba161b99c068036b8e in that repo).
This enables EPGs to be associated with multiple VMM Domains,
instead of just OpenStack VMM Domains, which is needed to support
ESX hypervisors.

It also fixes a few bugs in the DVS notifier code.

Change-Id: I088de6b8ce67a3c4e56883d7dc84a4ce1aa42055