[AIM] Validation/Repair/Migration Tool
Adds validation and repair framework that calls into mechanism, policy
and SFC drivers to validate mappings to Neutron resources and to AIM
resources. The mappings from all standard Neutron resources to AIM are
currently handled. New unit tests are provided for validation of each
resource, and validation calls are added to several existing unit
tests.

A simple command line interface is used perform validation, optionally
repairing invalid state. This is run using 'gbp-validate [--repair]
<neutron-server args>'. The same arguments that are passed to specify
the configuration of neutron-server must be passed to gbp-validate.

Validation of GBP and SFC resources and of SVI networks is not yet
implemented. Attempting to validate deployments where these resources
exist currently intentionally fails, even when repair is
enabled. Proper validation of these resources will be addressed in
followup patches.

For isomorphic address scopes with a non-pre-existing VRF, the VRF's
display name currently depends on the order in which the address
scopes were created. This will be addressed in a followup patch.

EPG domain association, static paths, and other aspects of port
binding are not yet validated. This will be addressed in a followup
patch.

Migration from the old APIC plugin to the unified plugin will require
associating existing subnets with subnetpools, which will also be
addressed in a followup patch.

Change-Id: I0c3fe9e2629f76ecca8b3c8a93f9534b2d946e14