Create notifications for dependent ports
Changes to certain neutron resources, such as DHCP or LBaaS ports,
require notiications be generated to ports who depend on those
changes (e.g. if a DHCP agent port state changes, the host route
to the metadata agent needs updating). This patch checks if:
   o  an update or delete to a DHCP or LBaaS port occurs
   o  the port belongs to an opflex type network

If both of the above conditions are met, the plugin generates port
update notifications to all the ports on that network, so they can
update their parameters using the request_endpoint_details_list RPC.

Change-Id: Ida893b1c331c4778a1276de54a379627d54b7d23