Add RPC for network ports
This adds an RPC to request notifications for ports on
a network. This will be used by a subsequent patch to
trigger port update notifications whenever resources
unrelated to a port change (e.g. let an instance's port
get notified if a HA DHCP port on a given network goes
down).

Change-Id: Ib316028dc9cfd8dd368787e7a25058af49f3e928