[WIP] SVI network support.
1. Introduce the apic:svi network extension attribute to indicate its a SVI network.
2. For SVI network, user can do either pre-existing l3out or auto-l3out (which means we
will create the whole l3out for the user). For pre-existing l3out, user has to provide
the external network DN while creating the SVI network.
3. When auto-l3out is being used, the l3out vrf/teant can be changed accordingly when this
SVI network is connected to a router. This will not happen for the pre-existing l3out mode
of course.
4. When VM is being spun on a host, the corresponding host link(s) will be used to create
the SVI interface under the l3out while for the non-SVI network, the host link(s) will be
added to the static paths of the corresponding EPG. There will be no EPG/BD being created
on the APIC side though for the SVI network as its using the l3out instead.
5. There are still some TODO's but ACI QA needs this to be merged soon as they are already
testing it. I've marked those in the code with TODO keyword.

Change-Id: I44e2d3ef57ba562c5f157c440f7f15d671536cf4
(cherry picked from commit 9bcfa3014edb28a1eaad183fd0418acf13e6a5fc)