Fix VMM domain EPG assignment
Commit 6bbbb6cb2266d059df4ee109c588fee5e0d5d0be introduced a
backwards-incompatible change to the assignment of VMM domains
to EPGs in APIC. If the Host Domain Mapping table had no entries,
then the default behavior was to only assign OpenStack VMM Domains
to the EPGs. The above commit changed it such that both OpenStack
and VMware VMM Domains got assigned to EPGs. This commit reverts
that behavior, ensuring that only OpenStack VMM Domains are
assigned when the table is empty.

Change-Id: I5e610daa025c3992f841ae5fcae6f0cca3776880