Query DB to get the tenant ID of the SG
There is a use case that as an admin user, people can just issue
port-udpate command to set the SG even if that SG is from another tenant.
In this use case, the tenant ID of the SG could be from any tenant so we
have to do a DB query to get that.

Change-Id: I8b52b049928f1a0b6db56749c02a13cab39a2f98
(cherry picked from commit 7337427275c1b6e1618a720c74c1021211f07c21)