Patch writer context in trunk plugin
Neutron partially migrated to the reader/writer pattern for starting
a transaction in the Ocata release. The trunk plugin was one of the
modules that got upgraded. However, the old and new patterns don't
always mix well. This is especially true in the way that the apic_aim
mechanism driver in GBP uses it. GBP moved to using the new pattern in
Pike (at the point that Neutron finished its migration). But the new
pattern used in the trunk plugin creates a problem with transaction
consistency in Ocata. Hence its being patched here to revert to using
the older pattern.

Change-Id: I643c8fa510f83e88de2307b96668c6280141e5ac