[APIC AIM] Ignore reverse network mapping errors
Reverse network mapping is won't work for BDs in APIC
that the mechanism driver is unaware of, for example
pre-existing VRFs that the user has created. So
we avoid raising an exception in get_gbp_details
if we cannot resolve a BD found in APIC to a Neutron
network.
I've deliberatedly not added a log message to avoid
spamming the log stream.

Change-Id: I9f1a1f44051e75956f3cad8af50774b10d6316f7
Signed-off-by: Amit Bose <amitbose@gmail.com>