[AIM] Avoid link update race conditions
When moving compute nodes' cables around, we receive
from the LLDP agent a link_update or link_delete
notification. This would often result in leaked
or missing configuration due to multiple possible race
conditions that the old implementation was subject to.
By regenerating the configuration by host, together
with the AIM object versioning implementation and the
usage of the DB retry decorator, we can avoid such
rage conditions and guarantee that the system will
eventually be consistent.

Change-Id: I4e3d94a1fc4f73f54d92a109d6d08a6847eeb92c