Split the default ext epg 0/0 subnet into 0/1 and 128/1
Since each SVI ext epg subnet is 0/0 so if there are multiple SVI networks
under this VRF then the system doesn't know which contract to use. We still
need 0/0 in each SVI so that BGP can work properly but for each VRF we will
create a separate default SVI l3out that will have subnets 0/1 and 128/1.
We will create this default SVI l3out when the first SVI l3out is created
under this VRF. Vice versa We will delete this default SVI l3out when the
last SVI l3out is being deleted under this VRF. When a SVI network is
connected to a router then we will apply the router contract to this default
SVI ext epg instead. This way there will be no ambiguity regarding which
contract to use when traffic comes in.

Change-Id: I7d7a737664d0cdeb88b0197e12215a804a174d1b