Reduce the DBDuplicateEntry errors while creating a SVI subnet
This error would happen because when the subnet is getting created,
3 DHCP ports are also being created all at once as there are 3
controllers in the setup. This means there will be 3 threads trying
to create the SVI interface concurrently.

Also put a check to make sure the subnet exists for the network before
we proceed further while creating the SVI interface.

Change-Id: I87409d32fccf8f9f61e54d998b38f162fdb1d9e3