Fix isolated subnets condition in gbp_details
The get_gbp_details provides static routes to the agent,
which it uses in its DHCP offers when optimized DHCP is
turned on. If a subnet isn't connected to a neutron router,
then a host route is required for the metadata service via
the DHCP agent. The current check for isolated subnets is
incorrect, which is fixed in this patch.

Change-Id: I1d7aeee8574d7cc50aa8ebc3d779392c136ef18f